Alterado a tela que era a MainActivity para a nova com os menus.
