+++ app/src/main/AndroidManifest.xml
@@ -47,7 +47,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".activity.NavigatorMenu" />
+        <activity android:name=".activity.MainActivity" />
         <activity
             android:name=".activity.CadastroPetActivity"
             android:windowSoftInputMode="stateHidden" />
