+++ app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 app:layout_behavior="@string/appbar_scrolling_view_behavior"
-tools:context="projetaobcc20172.com.projetopetemfoco.activity.NavigatorMenu"
+tools:context="projetaobcc20172.com.projetopetemfoco.activity.MainActivity"
 tools:showIn="@layout/app_bar_main">
